[TAN-7616] Clean up background_jobs route in frontend
