+++ front/app/components/admin/PostManager/CommonGroundInputManager/ImportInputsModal.tsx
@@ -4,7 +4,7 @@ import { Box, Button, Text } from '@citizenlab/cl2-component-library';
 import { useParams } from 'react-router-dom';
 import { IOption } from 'typings';
 
-import { IJob } from 'api/copy_inputs/types';
+import { IJob } from 'api/background_jobs/types';
 import useCopyInputs from 'api/copy_inputs/useCopyInputs';
 import usePhase from 'api/phases/usePhase';
